moved methods from inspector tool to respective module_errors.rb file
